external modules: directory and excel etl refactoring
